Some route path changes Safe did upstream (#60)
* Update securityHeaders.ts

* Update securityHeaders.ts

* Update wallets.ts

* Update wallets.ts

* Update sdk.ts

* Update useWalletCanRelay.ts

* Update hooks.ts

* Update hooks.test.ts

* Update SignOrExecuteForm.test.tsx

* Update consts.ts

* Update consts.ts

* Update action.yml

* Update securityHeaders.ts

* Update securityHeaders.ts

* Update securityHeaders.ts
