+++ src/components/tx/SignOrExecuteForm/SignOrExecuteForm.test.tsx
@@ -7,7 +7,7 @@ import * as useChainsHook from '@/hooks/useChains'
 import * as txSenderDispatch from '@/services/tx/tx-sender/dispatch'
 import * as wallet from '@/hooks/wallets/useWallet'
 import * as onboard from '@/hooks/wallets/useOnboard'
-import * as walletUtils from '@/hooks/wallets/wallets'
+import * as walletUtils from '@/utils/wallets'
 import * as web3 from '@/hooks/wallets/web3'
 import type { ChainInfo, SafeInfo, TransactionDetails } from '@safe-global/safe-gateway-typescript-sdk'
 import { waitFor } from '@testing-library/react'
