+++ src/hooks/useWalletCanRelay.ts
@@ -1,7 +1,7 @@
 import useAsync from '@/hooks/useAsync'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import useWallet from '@/hooks/wallets/useWallet'
-import { isSmartContractWallet } from '@/hooks/wallets/wallets'
+import { isSmartContractWallet } from '@/utils/wallets'
 import { Errors, logError } from '@/services/exceptions'
 import { hasEnoughSignatures } from '@/utils/transactions'
 import { type SafeTransaction } from '@safe-global/safe-core-sdk-types'
